+++ backend/src/main/java/com/grepp/smartwatcha/app/model/admin/user/dto/AdminSimpleRatingDto.java
@@ -1,4 +1,4 @@
-package com.grepp.smartwatcha.app.model.admin.user;
+package com.grepp.smartwatcha.app.model.admin.user.dto;
 
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
