+++ backend/src/main/java/com/grepp/smartwatcha/app/model/admin/user/repository/AdminUserRatingJpaRepository.java
@@ -1,4 +1,4 @@
-package com.grepp.smartwatcha.app.model.admin.user;
+package com.grepp.smartwatcha.app.model.admin.user.repository;
 
 import com.grepp.smartwatcha.infra.jpa.entity.RatingEntity;
 import java.util.List;
